Renamed combinator module.
