+++ test/itest_break_reverse.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import testlib
-import testlib_combinatoric_tests
+import test_combinators
 
 fee = 20
 initialsend = 200000
@@ -10,7 +10,7 @@ capacity = 1000000
 def run_test(env):
     lit1 = env.lits[0]
     lit2 = env.lits[1]
-    testlib_combinatoric_tests.run_break_test(env, lit1, lit2, lit1)
+    test_combinators.run_break_test(env, lit1, lit2, lit1)
 
 if __name__ == '__main__':
     env = None
