+++ test/itest_pushclose_reverse.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import testlib
-import testlib_combinatoric_tests
+import test_combinators
 
 fee = 20
 initialsend = 200000
@@ -12,7 +12,7 @@ pushsend = 250000
 def run_test(env):
     lit1 = env.lits[0]
     lit2 = env.lits[1]
-    testlib_combinatoric_tests.run_pushclose_test(env, lit1, lit2, lit2)
+    test_combinators.run_pushclose_test(env, lit1, lit2, lit2)
 
 if __name__ == '__main__':
     env = None
